Fight simulation in frontend

+++ duck-family-api-lib/src/attributes.rs
@@ -2,8 +2,8 @@ use crate::types::*;
 
 pub trait Attributes {
     fn range(&self) -> Option<f32>;
-    fn attack_power(&self) -> Option<f32>;
-    fn attacks_per_cycle(&self) -> Option<i32>;
+    fn attack_power(&self) -> Option<i64>;
+    fn attacks_per_cycle(&self) -> Option<i64>;
     fn size(&self) -> (usize,usize);
 }
 
@@ -16,13 +16,13 @@ impl Attributes for BuildingType {
             BuildingType::RedFlowers => Some(1.0)
         }
     }
-    fn attack_power(&self) -> Option<f32> {
+    fn attack_power(&self) -> Option<i64> {
         match self {
-            BuildingType::BlueFlowers => Some(1.0),
-            BuildingType::RedFlowers => Some(3.0)
+            BuildingType::BlueFlowers => Some(1),
+            BuildingType::RedFlowers => Some(3)
         }
     }
-    fn attacks_per_cycle(&self) -> Option<i32>{
+    fn attacks_per_cycle(&self) -> Option<i64>{
         match self {
             BuildingType::BlueFlowers => None,
             BuildingType::RedFlowers => None
